requirements.txt
numpy==1.22.4
pandas==1.4.2
plotly==5.8.0
matplotlib==3.5.2
typing_extensions==4.2.0
pip==22.1.1
scikit-learn==1.1.1
mitosheet==0.1.426
streamlit-option-menu==0.3.2